serviceReservation/src/main/java/ma/emsi/serviceReservation/model/Ville.java
package ma.emsi.serviceReservation.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

public class Ville {
	private int id;
	private String label;

	public Ville() {
		super();
	}

	public Ville(int id, String label) {
		super();
		this.id = id;
		this.label = label;
	}

	public int getId() {
		return id;
	}

	public void setId(int id) {
		this.id = id;
	}

	public String getLabel() {
		return label;
	}

	public void setLabel(String label) {
		this.label = label;
	}

	@Override
	public String toString() {
		return "Ville [id=" + id + ", label=" + label + "]";
	}

}
